+++ frontend/app/src/components/bots/AutoBotBuilder.svelte
@@ -143,6 +143,7 @@
             autonomousConfig: undefined,
             defaultSubscriptions: undefined,
             dataEncoding: undefined,
+            restrictedLocations: undefined,
         };
     }